Temporary commit to check for local issues.

+++ user-microservice/src/main/java/nl/tudelft/sem/template/domain/user/UserService.java
@@ -16,7 +16,7 @@ public class UserService {
      * @param userId - id of a user
      * @return - user with this id if exists, else null
      */
-    public AppUser getUserById(int userId) {
+    public AppUser getUserById(long userId) {
         if (userRepository.findById(String.valueOf(userId)).isPresent()) {
             return userRepository.findById(String.valueOf(userId)).get();
         }
